backend/pyproject.toml
[project]
name = "libraryweb"
version = "0.1.3"
description = ""
authors = [
    {name = "dezolute"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyjwt (>=2.10.1,<3.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "fastapi (>=0.116.1,<0.117.0)",
    "alembic (>=1.16.5,<2.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "pydantic[email] (>=2.11.9,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "black (>=25.1.0,<26.0.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "pytest (>=8.4.2,<9.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "boto3 (>=1.40.45,<2.0.0)",
    "sqladmin[full] (>=0.21.0,<0.22.0)",
    "aiohttp (>=3.13.0,<4.0.0)",
    "redis (>=6.4.0,<7.0.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
